# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Port Logistics API",
        "description": "Port Logistics API",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://port_logistic.com/",
            "description": ""
        }
    ],
    "paths": {
...
        "/consignee/{id}": {
            "get": {
                "tags": [
                    "Consignee"
                ],
                "summary": "Consignee",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ConsigneeController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of consignee",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Consignee"
                ],
                "summary": "Consignee",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ConsigneeController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of consignee",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Update consignee details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Tata Power"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Tata power"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Consignee"
                ],
                "summary": "Consignee",
                "description": "Delete consignee details",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ConsigneeController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of consignee",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/department": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get list of departments",
                "description": "Returns list of departments",
                "operationId": "App\\Http\\Controllers\\API\\V1\\DepartmentController::getAllDepartments",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Department"
                ],
                "summary": "Department",
                "operationId": "App\\Http\\Controllers\\API\\V1\\DepartmentController::storeDepartments",
                "requestBody": {
                    "description": "Pass department details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Management"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Management"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/department/{id}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Department",
                "description": "Edit department details",
                "operationId": "App\\Http\\Controllers\\API\\V1\\DepartmentController::editDepartments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of department",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Department"
                ],
                "summary": "Department",
                "description": "Update cargo details",
                "operationId": "App\\Http\\Controllers\\API\\V1\\DepartmentController::updateDepartments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of department",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "update department details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Development"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Developers Dept"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Department"
                ],
                "summary": "Department",
                "description": "Delete department details",
                "operationId": "App\\Http\\Controllers\\API\\V1\\DepartmentController::deleteDepartments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of department",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and mobile to get the authentication token",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}